+++ src/index.js
@@ -3,6 +3,7 @@
 export { TreeShaker } from './tree-shaking.js';
 export { Minifier } from './minification.js';
 export { CSSMinifier } from './css-minification.js';
+export { HTMLMinifier } from './html-minification.js';
 export { Bundler } from './bundling.js';
 export { ModuleSystem } from './modules.js';
 export { Obfuscator } from './obfuscation.js';
@@ -14,6 +15,7 @@ if (typeof module !== 'undefined' && module.exports) {
     TreeShaker: require('./tree-shaking.js').TreeShaker,
     Minifier: require('./minification.js').Minifier,
     CSSMinifier: require('./css-minification.js').CSSMinifier,
+    HTMLMinifier: require('./html-minification.js').HTMLMinifier,
     Bundler: require('./bundling.js').Bundler,
     ModuleSystem: require('./modules.js').ModuleSystem,
     Obfuscator: require('./obfuscation.js').Obfuscator,